Fix turning LEDs off on suspend
When suspending, we want to set all LEDs off & sync first, and only then pause
LED operations. When the code was originally written, `LEDControl.syncLeds()`
synced LEDs even when `LEDControl.paused` was set - it no longer does so, so we
need to pause after we synced.

This addresses keyboardio/Kaleidoscope#516, at least partially.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>